requirements.txt
boto3==1.6.6
SpeechRecognition==3.8.1
picamer==1.13
